requirements.txt
OWSLib==0.12.0
Pillow==3.3.0
